requirements.txt
# Streamlit Cloud deployment dependencies
streamlit>=1.40.0
plotly>=5.24.0
python-dotenv>=1.1.0

# TradingAgents core
langchain-core>=0.3.81
langchain-anthropic>=0.3.15
langchain-experimental>=0.3.4
langchain-google-genai>=2.1.5
langchain-openai>=0.3.23
langgraph>=0.4.8
pandas>=2.3.0
yfinance>=0.2.63
requests>=2.32.4
rank-bm25>=0.2.2
numpy
